+++ src/main/java/com/zycus/boot/services/UserService.java
@@ -18,7 +18,7 @@ public class UserService {
 		return userRepository.findAll();
 	}
 	
-	public User getUserById(Long id)
+	public User getUserById(Integer id)
 	{
 		return userRepository.findById(id).get();
 	}
